hand-made dataset

+++ project/src/network_math.rs
@@ -1,4 +1,4 @@
-pub fn product(arr: &Vec<Vec<f64>>, vec: &[f64], result: &mut Vec<f64>) {
+pub fn product(arr: &Vec<Vec<f32>>, vec: &[f32], result: &mut Vec<f32>) {
     //todo improve it, make it parallel or use a library for matrix product
     for i in 0..arr.len() {
         let mut tmp = 0.0;
@@ -9,7 +9,7 @@ pub fn product(arr: &Vec<Vec<f64>>, vec: &[f64], result: &mut Vec<f64>) {
     }
 }
 
-pub fn sum(vec1_result: &mut Vec<f64>, vec2: &Vec<f64>) {
+pub fn sum(vec1_result: &mut Vec<f32>, vec2: &Vec<f32>) {
     //todo improve it, make it parallel or use a library for matrix product
     for i in 0..vec1_result.len() {
         vec1_result[i] = vec1_result[i] + vec2[i];
